refactor: Move admin-config service to components
Move the `admin-config` module from `services` to `components` to align
with the new module structure.
