+++ components/admin-config/backend/src/main/kotlin/ScannerConfig.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.services.config
+package org.eclipse.apoapsis.ortserver.components.adminconfig
 
 import org.eclipse.apoapsis.ortserver.model.SourceCodeOrigin
