+++ components/admin-config/backend/src/test/kotlin/ReporterConfigTest.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.eclipse.apoapsis.ortserver.services.config
+package org.eclipse.apoapsis.ortserver.components.adminconfig
 
 import io.kotest.core.spec.style.WordSpec
 import io.kotest.matchers.nulls.beNull
