+++ settings.gradle.kts
@@ -28,6 +28,7 @@ include(":api:v1:mapping")
 include(":api:v1:model")
 include(":cli")
 include(":clients:keycloak")
+include(":components:admin-config:backend")
 include(":components:authorization:api-model")
 include(":components:authorization:backend")
 include(":components:authorization-keycloak:backend")
@@ -71,7 +72,6 @@ include(":secrets:file")
 include(":secrets:spi")
 include(":secrets:scaleway")
 include(":secrets:vault")
-include(":services:admin-config")
 include(":services:hierarchy")
 include(":services:ort-run")
 include(":services:report-storage")
