+++ workers/config/build.gradle.kts
@@ -47,11 +47,11 @@ repositories {
 }
 
 dependencies {
+    implementation(projects.components.adminConfig.adminConfigBackend)
     implementation(projects.components.pluginManager.pluginManagerBackend)
     implementation(projects.config.configSpi)
     implementation(projects.dao)
     implementation(projects.model)
-    implementation(projects.services.adminConfigService)
     implementation(projects.transport.transportSpi)
     implementation(projects.utils.logging)
     implementation(projects.workers.common)
