+++ workers/notifier/src/test/kotlin/NotifierRunnerTest.kt
@@ -31,16 +31,16 @@ import io.mockk.verify
 
 import java.io.File
 
+import org.eclipse.apoapsis.ortserver.components.adminconfig.AdminConfig
+import org.eclipse.apoapsis.ortserver.components.adminconfig.AdminConfigService
+import org.eclipse.apoapsis.ortserver.components.adminconfig.JiraRestClientConfiguration
+import org.eclipse.apoapsis.ortserver.components.adminconfig.MailServerConfiguration
+import org.eclipse.apoapsis.ortserver.components.adminconfig.NotifierConfig
+import org.eclipse.apoapsis.ortserver.components.adminconfig.RuleSet
 import org.eclipse.apoapsis.ortserver.config.ConfigManager
 import org.eclipse.apoapsis.ortserver.config.Context
 import org.eclipse.apoapsis.ortserver.config.Path
 import org.eclipse.apoapsis.ortserver.model.JobConfigurations
-import org.eclipse.apoapsis.ortserver.services.config.AdminConfig
-import org.eclipse.apoapsis.ortserver.services.config.AdminConfigService
-import org.eclipse.apoapsis.ortserver.services.config.JiraRestClientConfiguration
-import org.eclipse.apoapsis.ortserver.services.config.MailServerConfiguration
-import org.eclipse.apoapsis.ortserver.services.config.NotifierConfig
-import org.eclipse.apoapsis.ortserver.services.config.RuleSet
 import org.eclipse.apoapsis.ortserver.workers.common.context.WorkerContext
 
 import org.ossreviewtoolkit.model.OrtResult
